+++ transducers/entry.lisp
@@ -118,6 +118,7 @@ streamed as-is as cons cells."
 
 (declaim (ftype (function ((function (&optional t t) *) t list) *) list-reduce))
 (defun list-reduce (f identity lst)
+  (declare (optimize (speed 3)))
   (labels ((recurse (acc items)
              (if (null items)
                  acc
@@ -139,6 +140,7 @@ streamed as-is as cons cells."
     (funcall xf result)))
 
 (defun vector-reduce (f identity vec)
+  (declare (optimize (speed 3)))
   (let ((len (length vec)))
     (labels ((recurse (acc i)
                (if (= i len)
@@ -159,6 +161,7 @@ streamed as-is as cons cells."
     (funcall xf result)))
 
 (defun reversed-reduce (f identity rev)
+  (declare (optimize (speed 3)))
   (let* ((vec (reversed-vector rev))
          (len (length vec)))
     (labels ((recurse (acc i)
@@ -189,6 +192,7 @@ streamed as-is as cons cells."
     (funcall xf result)))
 
 (defun hash-table-reduce (f identity ht)
+  (declare (optimize (speed 3)))
   (with-hash-table-iterator (iter ht)
     (labels ((recurse (acc)
                (multiple-value-bind (entry-p key value) (iter)
@@ -230,6 +234,7 @@ responsiblity of the caller!"
     (funcall xf result)))
 
 (defun stream-reduce (f identity stream)
+  (declare (optimize (speed 3)))
   (labels ((recurse (acc)
              (let ((line (read-line stream nil)))
                (if (not line)
@@ -252,6 +257,7 @@ responsiblity of the caller!"
     (funcall xf result)))
 
 (defun generator-reduce (f identity gen)
+  (declare (optimize (speed 3)))
   (labels ((recurse (acc)
              (let ((val (funcall (generator-func gen))))
                (cond ((eq *done* val) acc)
@@ -270,6 +276,7 @@ responsiblity of the caller!"
 
 (declaim (ftype (function ((function (&optional t t) *) t plist) *) plist-reduce))
 (defun plist-reduce (f identity lst)
+  (declare (optimize (speed 3)))
   (labels ((recurse (acc items)
              (cond ((null items) acc)
                    ((null (cdr items))